+++ docs/changelog.md
@@ -11,6 +11,7 @@
 - Support serializing indexes to zarr with compressed coordinates ({pull}`262`
 - Use `H3HexagonLayer` to visualize H3 data ({pull}`263`)
 - Support missing coordinates for the zarr convention and fail on missing required properties ({pull}`260`)
+- Infer a suitable convention if none given ({pull}`269`)
 
 ## 0.6.0 (2026-02-05)